Create  multiplication_table.py
Crash Course on Python
Practice Quiz: While Loops
Question 5:
The multiplication_table function prints the results of a number passed to it multiplied by 1 through 5. An additional requirement is that the result is not to exceed 25, which is done with the break statement. Fill in the blanks to complete the function to satisfy these conditions.